use float

+++ c++/utils/matrix/matrix.h
@@ -23,8 +23,8 @@ struct Shape {
     }
 };
 
-typedef double DATATYPE;
-// typedef float DATATYPE;
+// typedef double DATATYPE;
+typedef float DATATYPE;
 
 class Matrix {
